requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
branca==0.6.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.1
daphne==4.0.0
Django==4.2.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-redis==5.3.0
django-timezone-field==6.0
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.1
MarkupSafe==2.1.3
msgpack==1.0.5
Pillow==9.5.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.4
zope.interface==6.0
